+++ files/chef-marketplace-cookbooks/chef-marketplace/recipes/_compliance_enable.rb
@@ -1,5 +1,3 @@
-include_recipe 'chef-marketplace::_common_enable'
-
 directory '/etc/chef-compliance' do
   owner 'root'
   group 'root'