+++ files/chef-marketplace-ctl-commands/trim_actions_db.rb
@@ -1,6 +1,6 @@
 require 'marketplace/actions_trimmer'
 
-add_command_under_category 'trim-actions-db', 'Configuration', 'Trim the Chef Analytics actions database', 2 do
+add_command_under_category 'trim-actions-db', 'Maintenance', 'Trim the Chef Analytics actions database', 2 do
   @options = { db_size: 1, interval: 4 }
 
   OptionParser.new do |opts|